// src-ui/core/ipc/generated/chat/ChatAccountPreference.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChatTrayNotification } from "./ChatTrayNotification";

export type ChatAccountPreference = { botId: string, selfId: string, enabled: boolean, background: boolean, tray: boolean, trayNotification: ChatTrayNotification, ignoredGroups: Array<string>, hiddenGroups: Array<string>, notifyUnknownGroups: boolean, };