// src-ui/core/ipc/generated/chat/ChatTrayNavigation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChatArchiveConversation } from "./ChatArchiveConversation";

export type ChatTrayNavigation = { v: number, botId: string, selfId: string, conversation: ChatArchiveConversation, };